requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
smartapi-python>=1.3.0
logzero==1.7.0
websocket-client==1.6.4
pyotp==2.9.0
requests>=2.31.0
plotly>=5.17.0
python-dotenv>=1.0.0
pytz>=2024.1
streamlit-autorefresh>=0.0.5
psycopg2-binary
python-dateutil==2.8.2
cashfree-pg==4.5.1
Pillow==10.0.0

